+++ All.ui/miniprogram/pages/UI-Pages/ui.checkInDatePicker/ui.checkInDatePicker.js
@@ -5,19 +5,43 @@ Page({
    * 页面的初始数据
    */
   data: {
-    multiple: true
+    multiple: true,
+    value: [],
+    valueName: ''
   },
 
   /**
    * 生命周期函数--监听页面加载
    */
   onLoad(options) {
+    const s = new Date()
 
+    const start = new Date(this.clear(s))
+    const end = new Date(start.getTime() + (6 * 86400000))
+    this.data.value = [start, end].map(this.clear)
+    this.setData({
+      valueName: this.data.value.join('/'),
+    })
+  },
+
+  clear: function(s) {
+    const yy = s.getFullYear()
+    const mm = s.getMonth() + 1
+    const dd = s.getDate()
+    return [yy, mm, dd].join('-')
   },
 
   open: function(event) {
     const { multiple } = event.currentTarget.dataset
     this.setData({ multiple }, () => {
+      this.selectComponent('#checkInDatePicker').open()
+    })
+  },
+  openValue: function() {
+    this.setData({
+      value: this.data.value,
+      multiple: true
+    }, () => {
       this.selectComponent('#checkInDatePicker').open()
     })
   },